report only option

+++ app/services/morphosource/merge_biological_specimen_service.rb
@@ -1,7 +1,7 @@
 module Morphosource
-  class MergeBsoService
+  class MergeBiologicalSpecimenService
 
-    def self.call(merge_to=nil, merge_from=nil, delete_dup=true)
+    def self.call(merge_to=nil, merge_from=nil, delete_dup=true, report_only=false)
       bso_to = BiologicalSpecimen.find(merge_to)
       bso_from = BiologicalSpecimen.find(merge_from)
 		  ie_list = []
@@ -11,23 +11,17 @@ module Morphosource
         media_list << m.id
         ie = m.imaging_event
         ie_list << ie.id
-        ie.physical_object_id = [bso_to.id]
-
-# might have to send ie.save to a job??
-
-#        ie.save
+        ie.physical_object_id = [bso_to.id]        
+      	ie.save unless report_only
 		  end
 		  if media_list.present?
 				UpdateWorkIndexJob.perform_later(merge_to)
-				puts "moved media #{media_list}, IE #{ie_list} from specimen #{merge_from} to specimen #{merge_to}"
-			else
-        puts "no media found"		  
 		  end
-		  if delete_dup
+		  if delete_dup && !report_only
 		  	bso_from.destroy 
-		  	puts "specimen #{bso_from} destroyed"
+		  	puts " specimen #{merge_from} destroyed"
 		  end
-	  return nil
+		  return media_list, ie_list
     end
 
   end
